refactor: extract movement stats accumulator
