+++ src/stats/calculators/movement.rs
@@ -7,12 +7,6 @@ enum MovementSpeedBand {
     Supersonic,
 }
 
-const ALL_MOVEMENT_SPEED_BANDS: [MovementSpeedBand; 3] = [
-    MovementSpeedBand::Slow,
-    MovementSpeedBand::Boost,
-    MovementSpeedBand::Supersonic,
-];
-
 impl MovementSpeedBand {
     fn as_label_value(self) -> &'static str {
         match self {
@@ -21,38 +15,12 @@ impl MovementSpeedBand {
             Self::Supersonic => "supersonic",
         }
     }
-
-    fn as_label(self) -> StatLabel {
-        StatLabel::new("speed_band", self.as_label_value())
-    }
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 struct MovementClassification {
     speed_band: MovementSpeedBand,
     height_band: PlayerVerticalBand,
-}
-
-impl MovementClassification {
-    fn labels(self) -> [StatLabel; 2] {
-        [self.speed_band.as_label(), self.height_band.as_label()]
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Serialize, Deserialize, ts_rs::TS)]
-#[ts(export)]
-pub struct MovementStats {
-    pub tracked_time: f32,
-    pub total_distance: f32,
-    pub speed_integral: f32,
-    pub time_slow_speed: f32,
-    pub time_boost_speed: f32,
-    pub time_supersonic_speed: f32,
-    pub time_on_ground: f32,
-    pub time_low_air: f32,
-    pub time_high_air: f32,
-    #[serde(default, skip_serializing_if = "LabeledFloatSums::is_empty")]
-    pub labeled_tracked_time: LabeledFloatSums,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, ts_rs::TS)]
@@ -72,104 +40,11 @@ pub struct MovementEvent {
     pub height_band: String,
 }
 
-impl MovementStats {
-    pub fn average_speed(&self) -> f32 {
-        if self.tracked_time == 0.0 {
-            0.0
-        } else {
-            self.speed_integral / self.tracked_time
-        }
-    }
-
-    pub fn average_speed_pct(&self) -> f32 {
-        self.average_speed() * 100.0 / CAR_MAX_SPEED
-    }
-
-    pub fn slow_speed_pct(&self) -> f32 {
-        if self.tracked_time == 0.0 {
-            0.0
-        } else {
-            self.time_slow_speed * 100.0 / self.tracked_time
-        }
-    }
-
-    pub fn boost_speed_pct(&self) -> f32 {
-        if self.tracked_time == 0.0 {
-            0.0
-        } else {
-            self.time_boost_speed * 100.0 / self.tracked_time
-        }
-    }
-
-    pub fn supersonic_speed_pct(&self) -> f32 {
-        if self.tracked_time == 0.0 {
-            0.0
-        } else {
-            self.time_supersonic_speed * 100.0 / self.tracked_time
-        }
-    }
-
-    pub fn on_ground_pct(&self) -> f32 {
-        if self.tracked_time == 0.0 {
-            0.0
-        } else {
-            self.time_on_ground * 100.0 / self.tracked_time
-        }
-    }
-
-    pub fn low_air_pct(&self) -> f32 {
-        if self.tracked_time == 0.0 {
-            0.0
-        } else {
-            self.time_low_air * 100.0 / self.tracked_time
-        }
-    }
-
-    pub fn high_air_pct(&self) -> f32 {
-        if self.tracked_time == 0.0 {
-            0.0
-        } else {
-            self.time_high_air * 100.0 / self.tracked_time
-        }
-    }
-
-    pub fn tracked_time_with_labels(&self, labels: &[StatLabel]) -> f32 {
-        self.labeled_tracked_time.sum_matching(labels)
-    }
-
-    pub fn complete_labeled_tracked_time(&self) -> LabeledFloatSums {
-        let mut entries: Vec<_> = ALL_PLAYER_VERTICAL_BANDS
-            .into_iter()
-            .flat_map(|height_band| {
-                ALL_MOVEMENT_SPEED_BANDS.into_iter().map(move |speed_band| {
-                    let mut labels = vec![speed_band.as_label(), height_band.as_label()];
-                    labels.sort();
-                    LabeledFloatSumEntry {
-                        value: self.labeled_tracked_time.sum_exact(&labels),
-                        labels,
-                    }
-                })
-            })
-            .collect();
-
-        entries.sort_by(|left, right| left.labels.cmp(&right.labels));
-
-        LabeledFloatSums { entries }
-    }
-
-    pub fn with_complete_labeled_tracked_time(mut self) -> Self {
-        self.labeled_tracked_time = self.complete_labeled_tracked_time();
-        self
-    }
-}
-
 #[derive(Debug, Clone, Default)]
 pub struct MovementCalculator {
-    player_stats: HashMap<PlayerId, MovementStats>,
+    stats: MovementStatsAccumulator,
     player_teams: HashMap<PlayerId, bool>,
     previous_positions: HashMap<PlayerId, glam::Vec3>,
-    team_zero_stats: MovementStats,
-    team_one_stats: MovementStats,
     events: EventStream<MovementEvent>,
 }
 
@@ -179,15 +54,15 @@ impl MovementCalculator {
     }
 
     pub fn player_stats(&self) -> &HashMap<PlayerId, MovementStats> {
-        &self.player_stats
+        self.stats.player_stats()
     }
 
     pub fn team_zero_stats(&self) -> &MovementStats {
-        &self.team_zero_stats
+        self.stats.team_zero_stats()
     }
 
     pub fn team_one_stats(&self) -> &MovementStats {
-        &self.team_one_stats
+        self.stats.team_one_stats()
     }
 
     pub fn events(&self) -> &[MovementEvent] {
@@ -211,26 +86,6 @@ impl MovementCalculator {
             speed_band,
             height_band,
         }
-    }
-
-    fn apply_classification(
-        stats: &mut MovementStats,
-        classification: MovementClassification,
-        dt: f32,
-    ) {
-        match classification.speed_band {
-            MovementSpeedBand::Slow => stats.time_slow_speed += dt,
-            MovementSpeedBand::Boost => stats.time_boost_speed += dt,
-            MovementSpeedBand::Supersonic => stats.time_supersonic_speed += dt,
-        }
-
-        match classification.height_band {
-            PlayerVerticalBand::Ground => stats.time_on_ground += dt,
-            PlayerVerticalBand::LowAir => stats.time_low_air += dt,
-            PlayerVerticalBand::HighAir => stats.time_high_air += dt,
-        }
-
-        stats.labeled_tracked_time.add(classification.labels(), dt);
     }
 
     pub fn update(
@@ -258,22 +113,7 @@ impl MovementCalculator {
                 continue;
             };
             let speed = player.speed().unwrap_or(0.0);
-            let stats = self
-                .player_stats
-                .entry(player.player_id.clone())
-                .or_default();
-            let team_stats = if player.is_team_0 {
-                &mut self.team_zero_stats
-            } else {
-                &mut self.team_one_stats
-            };
-
             if live_play {
-                stats.tracked_time += frame.dt;
-                stats.speed_integral += speed * frame.dt;
-                team_stats.tracked_time += frame.dt;
-                team_stats.speed_integral += speed * frame.dt;
-
                 let distance = if let Some(previous_position) =
                     self.previous_positions.get(&player.player_id)
                 {
@@ -281,18 +121,12 @@ impl MovementCalculator {
                 } else {
                     0.0
                 };
-                if distance > 0.0 {
-                    stats.total_distance += distance;
-                    team_stats.total_distance += distance;
-                }
 
                 let height_band = vertical_state
                     .band_for_player(&player.player_id)
                     .unwrap_or_else(|| PlayerVerticalBand::from_height(position.z));
                 let classification = Self::classify_movement(speed, height_band);
-                Self::apply_classification(stats, classification, frame.dt);
-                Self::apply_classification(team_stats, classification, frame.dt);
-                self.events.push(MovementEvent {
+                let event = MovementEvent {
                     time: frame.time,
                     frame: frame.frame_number,
                     player: player.player_id.clone(),
@@ -303,7 +137,9 @@ impl MovementCalculator {
                     distance,
                     speed_band: classification.speed_band.as_label_value().to_owned(),
                     height_band: classification.height_band.as_label().value.to_owned(),
-                });
+                };
+                self.stats.apply_event(&event);
+                self.events.push(event);
             }
 
             self.previous_positions
